UserService/src/main/java/com/user/userservice/service/CustomUserDetailsService.java
package com.user.userservice.service;

import com.user.userservice.entity.User;
import com.user.userservice.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CustomUserDetailsService implements UserDetailsService {
    private final UserRepository userRepository;

    @Override
    public User loadUserByUsername(String email) throws UsernameNotFoundException {
        return userRepository.findByEmail(email)
                .filter(user -> user != null && user.isEnabled())
                .orElseThrow(() -> new UsernameNotFoundException("User Not found"));
    }
}
